Add pages & Update App.vue to get router-view pages
