+++ frontend/src/views/TodoPage.vue
@@ -1,0 +1,18 @@
+<template>
+  <header>
+    <router-link to="/ts-work">TS Work</router-link>
+    <h1>TODO Page!!</h1>
+  </header>
+</template>
+
+<script>
+export default {
+  name: "TodoPage"
+}
+</script>
+
+<style scoped>
+header {
+  text-align: center;
+}
+</style>
